restructure: dlt contracts
