--- contracts/.gitignore
@@ -1,3 +1,0 @@
-!/artifacts/*
-/.starknet*
-/py-*
